Fix -verify with remote exec
This previously only worked with worker mode
